oot_collision/_classes reorg
