+++ fast64_internal/oot/collision/exporter/__init__.py
@@ -1,0 +1,9 @@
+from .functions import exportCollisionCommon
+
+from .classes import (
+    OOTCollision,
+    OOTCameraData,
+    OOTCameraPosData,
+    OOTWaterBox,
+    OOTCrawlspaceData,
+)
