+++ fast64_internal/oot/collision/exporter/to_c/__init__.py
@@ -1,0 +1,1 @@
+from .collision import ootCollisionToC, exportCollisionToC
