+++ fast64_internal/oot/spline/properties.py
@@ -3,7 +3,7 @@ from bpy.props import EnumProperty, PointerProperty, StringProperty, IntProperty
 from bpy.utils import register_class, unregister_class
 from ...utility import prop_split
 from ..oot_utility import drawEnumWithCustom
-from ..oot_collision_classes import ootEnumCameraCrawlspaceSType
+from ..collision.constants import ootEnumCameraCrawlspaceSType
 from ..actor.properties import OOTActorHeaderProperty
 from ..scene.properties import OOTAlternateSceneHeaderProperty
